+++ common/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-
+    <!-- Min Sdk version for YubiKit is 19, so override is needed to avoid build errors with Common's min sdk version -->
     <uses-sdk tools:overrideLibrary="com.yubico.yubikit.android" />
 
     <application>
