+++ common/src/main/java/com/microsoft/identity/common/internal/providers/oauth2/WebViewAuthorizationFragment.java
@@ -266,12 +266,14 @@ public class WebViewAuthorizationFragment extends AuthorizationFragment {
         mWebView.setWebViewClient(webViewClient);
     }
 
-    // In the case that AADWebViewClient is using EoClientCertAuthChallengeHandler to handle CBA,
+    // For ClientCertAuthChallengeHandler within AADWebViewClient,
     // the YubiKitManager needs to stop discovering Usb devices upon fragment destroy.
     @Override
     public void onDestroy() {
         super.onDestroy();
-        mAADWebViewClient.stopYubiKitManagerUsbDiscovery();
+        if (mAADWebViewClient != null) {
+            mAADWebViewClient.stopYubiKitManagerUsbDiscovery();
+        }
     }
 
     /**
